+++ Changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## vNext
 ### Added
 - ApiKeys can be generated and used to start a testrun
+- Added OAuth Authentication for testruns
 
 ### Changed
 - Display an error if the scanario can't be saved